fix: resolve TODO items 353-356 (Skin triangleStart overflow, SHADER_MANIFEST linkage, shader memory leak, Texture fileName docs)
Agent-Logs-Url: https://github.com/NathanNatNat/wow.export.cpp/sessions/5365c436-9d1c-4e13-8808-1dd7da0bb65a

+++ src/js/3D/Skin.h
@@ -15,7 +15,7 @@ struct SubMesh {
 	uint16_t level;
 	uint16_t vertexStart;
 	uint16_t vertexCount;
-	uint16_t triangleStart;
+	uint32_t triangleStart;
 	uint16_t triangleCount;
 	uint16_t boneCount;
 	uint16_t boneStart;
